start adding some directory structure for api
