+++ api/index.js
@@ -1,0 +1,5 @@
+module.exports.register = (server) => {
+  require('./main').register(server);
+  require('./category').register(server);
+  require('./sign').register(server);
+};
